+++ t/workflow.t
@@ -11,7 +11,7 @@ eval "require DBI";
 if ( $@ ) {
     plan skip_all => 'DBI not installed';
 } else {
-    plan tests => 39;
+    plan tests => 42;
 }
 
 require_ok( 'Workflow' );
@@ -98,41 +98,54 @@ my @result_data   = ( 'INITIAL', $date );
     SomeObserver->clear_observations;
     $wf->execute_action( 'null' );
     my @observations = SomeObserver->get_observations;
-    is( scalar @observations, 6,
-        'Three observations sent on workflow execute to two observers' );
 
-    is( $observations[0]->[2], 'save',
-        'Save observation generated first to first observer' );
-    is( $observations[1]->[2], 'save',
-        'Save observation generated first to second observer' );
+use Data::Dumper;
+#warn Dumper     $observations[0];
 
-    is( $observations[2]->[0], 'class',
+    is( scalar @observations, 12,
+        'Six observations sent on workflow execute to two observers' );
+
+    is( $observations[0]->[2], 'startup',
+        'Startup observation generated first to first observer' );
+    is( $observations[1]->[2], 'startup',
+        'Startup observation generated first to second observer' );
+
+    is( $observations[2]->[2], 'run',
+        'Run observation generated to first observer' );
+
+    is( $observations[4]->[2], 'save',
+        'Save observation generated to first observer' );
+
+    is( $observations[6]->[0], 'class',
         'Observation sent to configured class observer first' );
-    is( $observations[2]->[2], 'execute',
+    is( $observations[6]->[2], 'completed',
         'Class observer sent the correct execute action' );
-    is( $observations[2]->[3], 'INITIAL',
+    is( $observations[6]->[3]->{state}, 'INITIAL',
         'Class observer sent the correct old state for execute' );
 
-    is( $observations[3]->[0], 'sub',
+    is( $observations[7]->[0], 'sub',
         'Observation sent to configured subroutine observer second' );
-    is( $observations[3]->[2], 'execute',
+    is( $observations[7]->[2], 'completed',
         'Subroutine observer sent the correct execute action' );
-    is( $observations[3]->[3], 'INITIAL',
+    is( $observations[7]->[3]->{state}, 'INITIAL',
         'Subroutine observer sent the correct old state for execute' );
 
-    is( $observations[4]->[0], 'class',
+    is( $observations[8]->[0], 'class',
         'Observation sent to configured class observer first' );
-    is( $observations[4]->[2], 'state change',
+    is( $observations[8]->[2], 'state change',
         'Class observer sent the correct state change action' );
-    is( $observations[4]->[3], 'INITIAL',
+    is( $observations[8]->[3]->{from}, 'INITIAL',
         'Class observer sent the correct old state for state change' );
 
-    is( $observations[5]->[0], 'sub',
+    is( $observations[9]->[0], 'sub',
         'Observation sent to configured subroutine observer second' );
-    is( $observations[5]->[2], 'state change',
+    is( $observations[9]->[2], 'state change',
         'Subroutine observer sent the correct state change action' );
-    is( $observations[5]->[3], 'INITIAL',
+    is( $observations[9]->[3]->{from}, 'INITIAL',
         'Subroutine observer sent the correct old state for state change' );
+
+    is( $observations[10]->[2], 'finalize',
+        'Finalize observation generated to first observer' );
 }
 
 {